+++ DervetApp/src/renderer/components/Wizard/Components/Technologies/ICE.vue
@@ -12,6 +12,13 @@
           :errorMessage="getErrorMsg('name')">
         </text-input>
 
+        <text-area
+          v-model="description"
+          :metadata="metadata.description"
+          :isInvalid="submitted && $v.description.$error"
+          :errorMessage="getErrorMsg('description')">
+        </text-area>
+
         <text-input
           v-model="numGenerators"
           :metadata="metadata.numGenerators"