+++ DervetApp/src/renderer/components/Wizard/Overview/ProjectConfiguration.vue
@@ -13,6 +13,13 @@
         :errorMessage="getErrorMsg('name')">
       </text-input>
 
+      <text-area
+        v-model="description"
+        :metadata="metadata.description"
+        :isInvalid="submitted && $v.description.$error"
+        :errorMessage="getErrorMsg('description')">
+      </text-area>
+
       <text-input
         v-model="startYear"
         :metadata="metadata.startYear"